+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+# 0.2.0 (2020-12-21)
+
+- **Breaking:** the `stdout_str` and `stderr_str` methods on `bossy::Output` and `bossy::Error` now use `bossy::Error` instead of `std::str::Utf8Error` as their error type.
+- Added `run_and_wait_for_str` and `run_and_wait_for_string` methods to `bossy::Command` to make it super easy to get stdout as a string.
+
 # 0.1.4 (2020-12-09)
 
 - Added convenience methods to `bossy::Command` for setting `bossy::Stdio::null`.